Update README.md
Task 1:
Czech language for the Login Customizer plugin

Problem Statement: The current version of the Login Customizer plugin in WordPress lacks support for the Czech language, presenting a significant challenge for users seeking a localized and user-friendly login experience. The absence of Czech language compatibility hinders the customization and personalization of login pages for users who primarily operate websites or applications in the Czech language. 

Objective: By the end of the development of this feature, the user will be able to switch the language from the Language Switcher Option at the bottom of the Login Page. This feature aims to allow users to customize their login pages in Czech, ensuring a user-friendly experience for both Czech-speaking administrators and visitors. The plugin seeks to improve the user experience by localizing the interface.
___________________________________________________________________________


Task 2:

Create 2 new templates
This feature allows the user to select built-in templates. Three templates have been created so far. You must create two new templates to support the WordPress customizer and login page.

___________________________________________________________________________